refactor(async_hooks): split resource scopes
