+++ crates/perry-ext-net/src/lib.rs
@@ -352,11 +352,9 @@ enum PendingNetEvent {
     Data(i64, Bytes),
     /// Peer half-closed (FIN received); public readable-side `end` event.
     End(i64),
-    /// A queued `socket.write` finished. `.1` is the completion token and
-    /// `.2` is the write error message when the write failed.
+    /// A queued `socket.write` finished with a completion token and optional error.
     WriteComplete(i64, u64, Option<String>),
-    /// Writable-side shutdown requested by `socket.end()`, distinct from FIN.
-    /// `.1` is the completion token and `.2` is the shutdown error message.
+    /// `socket.end()` writable shutdown with a completion token and optional error.
     ShutdownComplete(i64, u64, Option<String>),
     Close(i64),
     Error(i64, String),
